# services/users/api/openapi.yaml
openapi: 3.0.0
info:
  title: Users Service API
  version: 1.0.0
  description: Сервис управления пользователями (гибридный REST API)
servers:
  - url: https://api.example.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Local development
...
paths:
...
  /users/admins:
    get:
      summary: Получение списка администраторов с пагинацией
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema: 
            type: string
            enum: [createdAt:asc, createdAt:desc, email:asc, email:desc]
            default: "createdAt:desc"
        - name: status
          in: query
          schema: 
            type: string
            enum: [active, inactive, blocked, pending, all]
            default: all
        - name: search
          in: query
          schema: { type: string, maxLength: 100 }
          description: Поиск по email, username, firstName, lastName
      responses:
        '200':
          description: Список администраторов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/customers:
    get:
      summary: Получение списка заказчиков с пагинацией
      tags: [Admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - name: sort
          in: query
          schema: 
            type: string
            enum: [createdAt:asc, createdAt:desc, email:asc, email:desc]
            default: "createdAt:desc"
        - name: status
          in: query
          schema: 
            type: string
            enum: [active, inactive, blocked, pending, all]
            default: all
        - name: search
          in: query
          schema: { type: string, maxLength: 100 }
          description: Поиск по email, username, firstName, lastName
      responses:
        '200':
          description: Список заказчиков
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен от Auth Service
  responses:
...
    Unauthorized:
      description: Неавторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        username: { type: string }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        role: 
          type: string
          enum: [admin, customer, contractor]
          description: "admin - администратор, customer - заказчик, contractor - исполнитель"
        status: { type: string, enum: [active, inactive, blocked, pending] }
        rating: { type: number, minimum: 0, maximum: 5, nullable: true, description: "Рейтинг исполнителя (только для роли contractor)" }
        completedTickets: { type: integer, minimum: 0, nullable: true, description: "Количество выполненных заявок (только для роли contractor)" }
        activeTickets: { type: integer, minimum: 0, nullable: true, description: "Количество активных заявок (только для роли contractor)" }
        companyId: { type: string, format: uuid, nullable: true, description: "ID компании (для заказчиков и исполнителей)" }
        emailVerifiedAt: { type: string, format: date-time, nullable: true }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    UsersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        pagination:
          type: object
          properties:
            page: { type: integer }
            limit: { type: integer }
            total: { type: integer }
            pages: { type: integer }
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { type: string }
        timestamp: { type: string, format: date-time }